+++ ios/Plugin/Plugin.swift
@@ -56,6 +56,7 @@ public class DatePickerPlugin: CAPPlugin {
     //
     // sizes
     private var defaultButtonHeight: CGFloat = 50
+    private var defaultPaddingHeight: CGFloat = 0
     private var defaultTitleHeight: CGFloat = 50
     private var defaultSpacerHeight: CGFloat = 1
     
@@ -111,8 +112,9 @@ public class DatePickerPlugin: CAPPlugin {
         self.pickerButtonFontColor = self.call?.getString("buttonFontColor") ?? self.defaultPickerButtonFontColor
         self.pickerMergedDateAndTime = self.call?.getBool("mergedDateAndTime") ?? self.defaultPickerMergedDateAndTime
         
+        self.defaultPaddingHeight = (UIDevice.current.hasNotch ? 15 : 0)
         
-        self.alertSize = CGSize(width: self.bridge.viewController.view.bounds.size.width, height: 250 + self.defaultButtonHeight)
+        self.alertSize = CGSize(width: self.bridge.viewController.view.bounds.size.width, height: 250 + self.defaultButtonHeight + self.defaultPaddingHeight)
         
         if (self.pickerTheme == "dark" || self.pickerTheme == "legacyDark") {
             self.defaultDark()
@@ -136,6 +138,9 @@ public class DatePickerPlugin: CAPPlugin {
             let tz = TimeZone(identifier: pickerTimezone ?? "UTC")
             dateFormatter.timeZone = tz;
         }
+        if ((self.pickerLocale) != nil) {
+            dateFormatter.locale = Locale(identifier: self.pickerLocale!)
+        }
         return dateFormatter.string(from: date)
     }
     
@@ -145,6 +150,9 @@ public class DatePickerPlugin: CAPPlugin {
         if (pickerTimezone != nil) {
             let tz = TimeZone(identifier: pickerTimezone ?? "UTC")
             dateFormatter.timeZone = tz;
+        }
+        if ((self.pickerLocale) != nil) {
+            dateFormatter.locale = Locale(identifier: self.pickerLocale!)
         }
         guard let dt = dateFormatter.date(from: date) else {
             self.call?.reject("Failed to parse date")
@@ -169,7 +177,7 @@ public class DatePickerPlugin: CAPPlugin {
         if (self.doneButton == nil) {
             self.doneButton = UIButton(type: .custom)
         }
-        self.doneButton?.frame = CGRect(x: buttonWidth, y: self.alertSize.height - self.defaultButtonHeight, width: buttonWidth, height: self.defaultButtonHeight)
+        self.doneButton?.frame = CGRect(x: buttonWidth, y: self.alertSize.height - self.defaultButtonHeight - self.defaultPaddingHeight, width: buttonWidth, height: self.defaultButtonHeight)
         self.doneButton?.setTitle(self.pickerDoneText, for: .normal)
         self.doneButton?.setTitleColor(UIColor(hexString: self.pickerButtonFontColor), for: .normal)
         self.doneButton?.backgroundColor = UIColor(hexString: self.pickerButtonBgColor)
@@ -180,7 +188,7 @@ public class DatePickerPlugin: CAPPlugin {
     private func createCancelButton() {
         let buttonWidth =  self.alertSize.width / 2
         if (self.cancelButton == nil) {
-            self.cancelButton = UIButton(frame: CGRect(x: 0, y: self.alertSize.height - self.defaultButtonHeight, width: buttonWidth, height: self.defaultButtonHeight))
+            self.cancelButton = UIButton(frame: CGRect(x: 0, y: self.alertSize.height - self.defaultButtonHeight - self.defaultPaddingHeight, width: buttonWidth, height: self.defaultButtonHeight))
         }
         self.cancelButton?.setTitle(self.pickerCancelText, for: .normal)
         self.cancelButton?.setTitleColor(UIColor(hexString: self.pickerButtonFontColor), for: .normal)
@@ -265,7 +273,7 @@ public class DatePickerPlugin: CAPPlugin {
         
         self.alertView?.frame = CGRect(x: 0, y: 0, width: self.alertSize.width, height: self.alertSize.height)
         
-        let yPosition = (self.alertView?.bounds.size.height)! - self.defaultButtonHeight - self.defaultSpacerHeight
+        let yPosition = (self.alertView?.bounds.size.height)! - self.defaultButtonHeight - self.defaultSpacerHeight - self.defaultPaddingHeight
         let lineView = UIView(frame: CGRect(x: 0, y: yPosition, width: alertView!.bounds.size.width, height: self.defaultSpacerHeight))
         
         lineView.backgroundColor = UIColor(red: 198/255, green: 198/255, blue: 198/255, alpha: 1)
@@ -282,9 +290,9 @@ public class DatePickerPlugin: CAPPlugin {
     
     @objc func titleChange(_ date: Date) -> String{
         if (self.pickerTitle == nil) {
-            var format: String = "E, MMM d, yyyy HH:mm"
+            var format: String = self.picker24h ? "E, MMM d, yyyy HH:mm" : "E, MMM d, yyyy hh:mm a"
             if (self.pickerMode == "time") {
-                format = "HH:mm a"
+                format = self.picker24h ? "HH:mm" : "hh:mm a"
             } else if (self.pickerMode == "date") {
                 format = "E, MMM d, yyyy"
             }
@@ -368,6 +376,7 @@ public class DatePickerPlugin: CAPPlugin {
     
     @objc func darkMode(_ call: CAPPluginCall) {
         DispatchQueue.main.async {
+            
             let black = UIColor(hexString: "#121212")
             let white = UIColor(hexString: "#fafafa")
             self.alertView?.backgroundColor = black
@@ -432,6 +441,17 @@ public class DatePickerPlugin: CAPPlugin {
             UIView.animate(withDuration: 0.2, delay: 0, options: [.curveEaseInOut], animations: {
                 self.alertView!.center.y -= height
             }, completion: nil)
+        }
+    }
+}
+
+extension UIDevice {
+    var hasNotch: Bool {
+        guard #available(iOS 11.0, *), let window = UIApplication.shared.windows.filter({$0.isKeyWindow}).first else { return false }
+        if UIDevice.current.orientation.isPortrait {
+            return window.safeAreaInsets.top >= 44
+        } else {
+            return window.safeAreaInsets.left > 0 || window.safeAreaInsets.right > 0
         }
     }
 }